+++ R/hydro_metadata_imgw.R
@@ -19,7 +19,7 @@ hydro_metadata_imgw <- function(interval){
 
   base_url <- "https://dane.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_hydrologiczne/"
   
-  if (!httr::http_error(base_url)) {
+  if (httr::http_error(base_url)) {
     stop(call. = FALSE, 
          paste0("\nDownload failed. ",
                 "Check your internet connection or validate this url in your browser: ",